move plugin imports so they are present when we generate website docs
