+++ go/cmd/vtorc/cli/plugin_grpctmclient.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package cli
 
 // Imports and register the gRPC tabletmanager client
